Move wallets/client/context stuff into ../hooks
The distinction between context and hooks did not make sense.
