+++ pages/wallets/index.js
@@ -1,9 +1,11 @@
 import { getGetServerSideProps } from '@/api/ssrApollo'
 import { Button } from 'react-bootstrap'
-import { useWallets, useTemplates, DndProvider, KeyStatus, useWalletsLoading, useKeyError, useWalletsError } from '@/wallets/client/context'
+import {
+  DndProvider, KeyStatus, useWallets, useTemplates, useWalletsLoading, useKeyError, useWalletsError,
+  usePassphrasePrompt, useShowPassphrase, useSetWalletPriorities
+} from '@/wallets/client/hooks'
 import { WalletCard, WalletLayout, WalletLayoutHeader, WalletLayoutLink, WalletLayoutSubHeader } from '@/wallets/client/components'
 import styles from '@/styles/wallet.module.css'
-import { usePassphrasePrompt, useShowPassphrase, useSetWalletPriorities } from '@/wallets/client/hooks'
 import { WalletSearch } from '@/wallets/client/components/search'
 import { useMemo, useState } from 'react'
 import { walletDisplayName } from '@/wallets/lib/util'
