--- wallets/client/context/index.js
@@ -1,7 +1,0 @@
-import WalletsProvider from './provider'
-
-export * from './provider'
-export * from './dnd'
-export * from './reducer'
-
-export default WalletsProvider
