+++ wallets/client/hooks/index.js
@@ -1,3 +1,4 @@
+export * from './global'
 export * from './payment'
 export * from './image'
 export * from './indicator'
@@ -8,3 +9,4 @@ export * from './query'
 export * from './logger'
 export * from './diagnostics'
 export * from './passphrase'
+export * from './dnd'
